refactor: extract mcp client runner
