+++ cmd/nfrx-mcp/main_test.go
@@ -6,6 +6,8 @@ import (
 	"net/http/httptest"
 	"strings"
 	"testing"
+
+	"github.com/gaspardpetit/nfrx/internal/mcp"
 )
 
 func TestProbeProviderSetsAcceptHeader(t *testing.T) {
@@ -19,7 +21,7 @@ func TestProbeProviderSetsAcceptHeader(t *testing.T) {
 	}))
 	defer srv.Close()
 
-	if err := probeProvider(context.Background(), srv.URL); err != nil {
+	if err := mcp.ProbeProvider(context.Background(), srv.URL); err != nil {
 		t.Fatalf("probeProvider returned error: %v", err)
 	}
 }
@@ -31,7 +33,7 @@ func TestProbeProviderReturnsBodyOnError(t *testing.T) {
 		_, _ = w.Write([]byte(msg))
 	}))
 	defer srv.Close()
-	err := probeProvider(context.Background(), srv.URL)
+	err := mcp.ProbeProvider(context.Background(), srv.URL)
 	if err == nil || !strings.Contains(err.Error(), msg) {
 		t.Fatalf("expected error containing %q got %v", msg, err)
 	}
